% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_final_analysis_rmd.R
\name{build_final_analysis_rmd}
\alias{build_final_analysis_rmd}
\title{Build the final_analysis.Rmd file}
\usage{
build_final_analysis_rmd(path)
}
\arguments{
\item{path}{pass a path to the function}
}
\description{
This function will create a placeholder for your final analysis Rmd file
}
\examples{
build_final_analysis_rmd()
}
