% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_README.R
\name{build_README}
\alias{build_README}
\title{Build the README.md file}
\usage{
build_README(path)
}
\arguments{
\item{path}{The path for the new project being created.}
}
\value{
README document for the new project
}
\description{
This function will create a placeholder for your README file
}
