% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directories.R
\name{create_directories}
\alias{create_directories}
\title{Create the necessary file structure}
\usage{
create_directories(path)
}
\arguments{
\item{path}{pass a path to the function}
}
\description{
This function creates the necessary file structure for your analysis.
}
\examples{
create_directories()
}
