% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethics_export.R
\docType{data}
\name{ethics}
\alias{ethics}
\title{Ethics checklist}
\format{
simply a character vector with all the makings of the ethics checklist
from Deon
}
\source{
https://deon.drivendata.org
}
\usage{
ethics
}
\description{
Ethics Checklist from Deon
}
\keyword{datasets}
