% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ethics_export.R
\name{ethics_export}
\alias{ethics_export}
\title{Create an Rda for Ethics data}
\usage{
ethics_export()
}
\arguments{
\item{NA}{no parameters}
}
\description{
Creates the ethics checklist data file
}
\examples{
ethics_export()
}
