% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello_word.R
\name{hello_world}
\alias{hello_world}
\title{Reproducible Analysis}
\usage{
hello_world(path)
}
\arguments{
\item{path}{pass a path to the function}
}
\description{
This should allow you to create reproducible analysis in R
}
\examples{
hello_world()
}
\keyword{analysis}
\keyword{reproducible}
