% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_README.R
\name{rm_body}
\alias{rm_body}
\title{Create the body for the Final Rmd Doc}
\usage{
rm_body(path)
}
\arguments{
\item{path}{The path for the new project being created.}
}
\value{
body contents for README file
}
\description{
Create the body for the final Rmd Doc
}
