% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_final_analysis_rmd.R
\name{build_final_analysis_rmd}
\alias{build_final_analysis_rmd}
\title{Build the final_analysis.Rmd file}
\usage{
build_final_analysis_rmd(path)
}
\arguments{
\item{path}{The path for the new project being created.}
}
\value{
Final analysis document for the new project
}
\description{
This function will create a placeholder for your final analysis Rmd file
}
